requirements.txt
aiogram==3.15.0
python-dotenv==1.0.1
redis[hiredis]==5.2.1
groq==0.9.0
httpx==0.27.0
google-generativeai==0.8.3
